// web/src/api/generated/models/TransferScope.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TransferScope = {
    SHARED: 'SHARED',
    PERSONAL: 'PERSONAL'
} as const;
export type TransferScope = typeof TransferScope[keyof typeof TransferScope];


export function instanceOfTransferScope(value: any): boolean {
    for (const key in TransferScope) {
        if (Object.prototype.hasOwnProperty.call(TransferScope, key)) {
            if (TransferScope[key as keyof typeof TransferScope] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TransferScopeFromJSON(json: any): TransferScope {
    return TransferScopeFromJSONTyped(json, false);
}

export function TransferScopeFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransferScope {
    return json as TransferScope;
}

export function TransferScopeToJSON(value?: TransferScope | null): any {
    return value as any;
}

export function TransferScopeToJSONTyped(value: any, ignoreDiscriminator: boolean): TransferScope {
    return value as TransferScope;
}